+++ data/eligibility_flow.go
@@ -5,7 +5,7 @@ import (
 )
 
 type EligibilityFlow interface {
-	ProcessSubject(subject *Subject, comparisonTime time.Time, flowCounty string, age float64, timeSinceConviction int) map[int]*EligibilityInfo
+	ProcessSubject(subject *Subject, comparisonTime time.Time, flowCounty string, age int, timeSinceConviction int) map[int]*EligibilityInfo
 	ChecksRelatedCharges() (result bool)
 }
